Move MiddlewareVersion function to functions package
